requirements.txt
pygame==2.6.1
requests==2.32.3
python-dotenv==1.1.0
twitch-dl==3.1.0
rich==14.0.0
tqdm==4.67.1
moviepy==2.2.1
selenium==4.33.0
tiktok-uploader==1.1.1
dailymotion==0.2.5
google-api-python-client==2.170.0
google-auth-oauthlib==1.2.2
